[NavigationRailView] Update inset handling for Navigation Rail & Bottom Nav
Correctly apply only the top, bottom, and start padding to a `NavigationRailView`, moving the inset logic that was incorrectly moved to `NavigationBarView` back to `BottomNavigationView`.

Closes #2252
